simple authorization checking with Gates

+++ src/Http/Controllers/ContentController.php
@@ -6,6 +6,7 @@ use Fbollon\LaraCmsLite\Models\Content;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Route;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ContentController extends Controller
 {
@@ -16,6 +17,8 @@ class ContentController extends Controller
      */
     public function index()
     {
+        Gate::authorize('lara-cms-lite-manage');
+        
         $contentQuery = Content::query();
         $contentQuery->where('name', 'like', '%'.request('q').'%');
         $contentQuery->with('User');
@@ -31,6 +34,8 @@ class ContentController extends Controller
      */
     public function create()
     {
+        Gate::authorize('lara-cms-lite-manage');
+        
         $routes = $this->getRoutes();
         return view('lara-cms-lite::contents.create', compact('routes'));
     }
@@ -65,6 +70,8 @@ class ContentController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('lara-cms-lite-manage');
+        
         $newContent = $request->validate([
             'name'        => 'required|max:100',
             'description' => 'required',
@@ -88,6 +95,8 @@ class ContentController extends Controller
      */
     public function show($id)
     {
+        Gate::authorize('lara-cms-lite-manage');
+        
         $content = Content::with(config('lara-cms-lite.user.className'))->find($id);
         $mediaItems = null;
         return view('lara-cms-lite::contents.show', compact('content', 'mediaItems'));
@@ -101,6 +110,8 @@ class ContentController extends Controller
      */
     public function edit(Content $content)
     {
+        Gate::authorize('lara-cms-lite-manage');
+        
         $routes = $this->getRoutes();
         return view('lara-cms-lite::contents.edit', compact('content', 'routes'));
     }
@@ -114,6 +125,8 @@ class ContentController extends Controller
      */
     public function update(Request $request, Content $content)
     {
+        Gate::authorize('lara-cms-lite-manage');
+        
         $contentData = $request->validate([
             'name'        => 'required|max:100',
             'description' => 'required',
@@ -138,6 +151,8 @@ class ContentController extends Controller
      */
     public function destroy(Request $request, Content $content)
     {
+        Gate::authorize('lara-cms-lite-manage');
+        
         if ($request->get('content_id') == $content->id && $content->delete()) {
             return redirect()->route('contents.index');
         }
